instancetype: Extract IgnoreableInferenceError into infer sub package
This error is specific to the infer from volume logic and should reside
within that sub package.
