+++ pkg/instancetype/BUILD.bazel
@@ -36,7 +36,6 @@ go_library(
 go_test(
     name = "go_default_test",
     srcs = [
-        "errors_test.go",
         "instancetype_suite_test.go",
         "instancetype_test.go",
     ],
