+++ pkg/instancetype/infer/errors_test.go
@@ -17,7 +17,7 @@
  *
  */
 
-package instancetype_test
+package infer_test
 
 import (
 	"errors"
@@ -25,20 +25,20 @@ import (
 	. "github.com/onsi/ginkgo/v2"
 	. "github.com/onsi/gomega"
 
-	instancetype "kubevirt.io/kubevirt/pkg/instancetype/errors"
+	"kubevirt.io/kubevirt/pkg/instancetype/infer"
 )
 
 var _ = Describe("Instancetype errors", func() {
 	Context("IgnoreableInferenceError", func() {
 		It("Passes through error message of wrapped error", func() {
 			err := errors.New("test error")
-			ignoreableInferenceError := instancetype.NewIgnoreableInferenceError(err)
+			ignoreableInferenceError := infer.NewIgnoreableInferenceError(err)
 			Expect(ignoreableInferenceError.Error()).To(Equal(err.Error()))
 		})
 
 		It("Passes through wrapped error on unwrap", func() {
 			err := errors.New("test error")
-			ignoreableInferenceError := instancetype.NewIgnoreableInferenceError(err)
+			ignoreableInferenceError := infer.NewIgnoreableInferenceError(err)
 			Expect(errors.Unwrap(ignoreableInferenceError)).To(Equal(err))
 		})
 	})
