+++ pkg/instancetype/infer/infer_suite_test.go
@@ -1,0 +1,13 @@
+package infer_test
+
+import (
+	"testing"
+
+	. "github.com/onsi/ginkgo/v2"
+	. "github.com/onsi/gomega"
+)
+
+func TestInfer(t *testing.T) {
+	RegisterFailHandler(Fail)
+	RunSpecs(t, "Infer Suite")
+}
